flask_app/requirements.txt
Flask==2.2.5
paho-mqtt==1.6.1
flask_sqlalchemy==3.0.5
flask_cors==3.0.9
